requirements.txt
pandas
numpy
scipy
statsmodels
duckdb
pyarrow
scikit-learn
streamlit
pytest
python-dotenv
tabulate
pyyaml
